Fix Bug: the reScan of a file was not creating a new object for the reducer, but modifying the old one...
Refactor: the scanDbRef does not need to store the scanType, it is now contained in the FileProps itself. On the other hand, we now need the map to use the FileProps.id as key, since the objects are now cloned...
The id prop of the FileProps now return the relpath, which is the only id it can have. It was not used before.
Rename variables for better understanding and finding when refactoring.