Enhance performances on scan
Refactor: remove scan fileProps from reducer, because the copy of the maps at each file addition had a huge impact on performances with many files to scan. Move them to a database.
Add a resultReducer, that computes the identicals, modified, duplicated, newFiles and dfRefFiles lists with queries from database.
Change a lot the database api to allow queries on FilePropsDb, FileProps, FilePropsDbDuplicates and even generic queries with no specific structure.